Add second part: A DiceRollTotal is a database Entity that stores the number of simulations, total number of rolls as well as the number of occurences for each roll result for an identity of N dice, M dice faces.
Add DiceRollTotalRepository as a spring framework database that stores DiceRollTotal instances.

Extend DiceController with an endpoint "/dicerolltotal" which directly accesses a DiceRollTotal via their ID.

Extend DiceController again with a "/dicerolltotalpercentage" resource: a DiceRollTotalPercentage mimics a DiceRollTotal in most ways, but the occurences of individual roll results are instead represented with a double in form of a percentage.